refactor: change dir name from "extractor" to "extractors"; use module sys
