+++ src/aws/__init__.py
@@ -1,0 +1,3 @@
+from .clients import s3_client
+
+__all__ = ["s3_client"]
