+++ src/azure/__init__.py
@@ -1,0 +1,3 @@
+from .clients import get_ocr_client
+
+__all__ = ["get_ocr_client"]
