+++ src/core/__init__.py
@@ -1,0 +1,3 @@
+from .settings import settings
+
+__all__ = ["settings"]
